Add calibration UI, env config, CLI & tests
Introduce an interactive calibration workflow and developer guidance, plus supporting changes

- Add app/calibration.py: a Tkinter-based CalibrationApp with candidate generation, task library, camera/gesture loop and voting UI to iteratively tune camera/animation parameters.
- Add AGENTS.md: repository guidelines for structure, testing, style, and contribution workflow.
- Update app/config.py: load .env overrides, helper parsing functions, and initialize CAMERA_CONFIG and ANIMATION_CONFIG from environment variables.
- Improve app/main_window.py: richer status bar, keyboard shortcuts, transient action statuses, gesture handling robustness, and minor UI text/behavior fixes.
- camera_handler.py: return frame without unnecessary uint8 cast.
- main.py: add argparse-driven entrypoint and --calibrate flag to launch the calibration workflow; factor run logic into main/_run_application.
- Add tests/test_gesture_recognizer.py: unit tests for GestureRecognizer covering common gestures.

These changes add an interactive calibration flow, environment-driven configuration, CLI convenience, UI polish, and unit coverage for gesture recognition.

+++ camera_handler.py
@@ -124,7 +124,7 @@ class CameraHandler:
                 hand_coords = (control_point.x, control_point.y)
 
         return CameraOutput(
-            frame=frame.astype(np.uint8),
+            frame=frame,
             gesture=gesture,
             coords=hand_coords,
         )
